Move `paths` object to separate module to remove circular imports
