+++ src/components/AuthRoute/AuthRoute.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Redirect, Route, RouteProps } from "react-router-dom";
 import { useStore } from "../../Store";
-import { paths } from "../../Router";
+import paths from "../../paths";
 
 const AuthRoute: React.FC<RouteProps> = ({ children, ...rest }) => {
   const { apiToken } = useStore();
